requirements.txt
urllib3<=1.25.11
selenium>=3.141.0
Unidecode>=1.1.1
numpy>=1.18.4
beautifulsoup4>=4.9.1
lxml>=4.5.1
pysimplegui>=4.29.0
fake-useragent>=0.1.11
termcolor>=1.1.0
chromedriver-autoinstaller>=0.2.2
stopit>=1.1.2
colorama>=0.4.1
python-dateutil>=2.8.1
selenium-wire>=2.1.1
websockets>=8.1
wheel>=0.36.2
pyopenssl>=20.0.1
requests>=2.22.0
scipy>=1.4.1
tabulate>=0.8.9
pillow>=8.0.1
pywin32>=227; sys_platform == 'win32' or sys_platform == 'win64'
demjson>=2.2.4
